migrate:types to its dedicated file for configuration.tsx
